requirements.txt
discord.py
mysqlclient
sqlalchemy
sqlalchemy-utils